==> performance_analysis/03.0_sfa_analysis.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sfma import Data, Variable, SplineVariable, SplineGetter, SplinePriorGetter, UniformPrior, SFMAModel
import matplotlib.ticker as mticker

You have to run this script a couple different times: 
For the primary models of the paper, run the cells below this one six times with a different selected_cause_name specified each time:
    selected_cause_name = one of "All causes", "Cardiovascular diseases", "Maternal disorders", "Breast cancer", "Cervical cancer", "Ovarian cancer", "Uterine cancer"
    covariate_name = "sdi"
    measure_name = "dalys"
    version = "2spline"
    lineartails = True
    trimmed = True
    crossectional = False

For the sensitivity analyses, you'll want to run the script with variations to the other parameters six times each with a different selected_cause_name specified each time. Here are the options that are available at the moment (4/22/2025):
    covariate_name = "sdi" or "ldi"
    measure_name = "dalys"
    version = "2spline" or "3spline"
    crosssectional = True or False

selected_cause_name = "Uterine cancer"
covariate_name = "sdi"
cause_subset = "top_and_sexspecific"
measure_name = "dalys"
version = "2spline"
lineartails = True
pct_trim = 0.05
crosssectional = False

if covariate_name != "sdi":
    version = version + "_" + covariate_name

if crosssectional:
    version = version + "_xsectional"

if lineartails == False:
    version = version + "_nonlineartails"

if pct_trim not in [0.05, 0]:
    version = version + "_" + str(round(pct_trim*100)) + "trim"
elif pct_trim == 0:
    version = version + "_notrim"
    
print(version)

cause_label = selected_cause_name.replace(" ", "_").lower()

data_path = "filepath/" + covariate_name + "_" + cause_subset + "/" + measure_name + "_" + cause_label + "_compiled.csv"
output_dir = "filepath/"

# Load the data
data = pd.read_csv(data_path)
data.head()

# define important variable names:
x = 'mean_predictor' # SDI, main predictor + spline variable
y = 'val' # outcome variable

# calculate the standard error of the outcome variable
data['log_se'] = (np.log(data['upper']) - np.log(data['lower'])) / (2 * 1.96)

# log outcome
data['log_val'] = np.log(data[y])

# Flip the outcome variable
data['neg_log_val'] = -data['log_val']

data.head()

# Setting model parameters
if covariate_name == "ldi": 
    data[x] = np.log(data[x])

# Number of knots
if "2spline" in version: 
    knots_version = [0, 0.33, 0.66, 1]
    degree_version = 2
elif "3spline" in version:
    knots_version = [0, 0.25, 0.5, 0.75, 1]
    degree_version = 3
else:
    raise ValueError("Invalid spline option specified.")

# type of tails
if "nonlineartails" in version: 
    tail_version = False
else:
    tail_version = True

if "xsectional" in version: 
    data = data[data['year_id'] == 2023]

nlocations = data['location_name'].nunique()
nyears = data['year_id'].nunique()
nages = data['age_group_name'].nunique()
print(f"Number of locations: {nlocations}")
print(f"Number of years: {nyears}")
print(f"Number of ages: {nages}")

# Group by 'year_id' and 'location_name' and check the size of each group
size_we_want = data['cause_name'].nunique()
group_sizes = data.groupby(['year_id', 'location_name']).size()

# Confirm that all groups have exactly 6 rows
if (group_sizes == size_we_want).all():
    print("Number of outcomes we want: " + str(size_we_want))
    print("All combinations of year_id and location_name have exactly the number of rows per outcome we want.")
else:
    print("Some combinations of year_id and location_name do not have a row for each outcome.")
    print(group_sizes[group_sizes != size_we_want])  # Print the problematic groups

num_rows = data.shape[0]
print(f"Number of rows in the dataset: {num_rows}")

# create model object with log transformed data
model_data = Data(obs = 'neg_log_val',
            obs_se = 'log_se'
            )

variables = [
    SplineVariable(x,
                   spline=SplineGetter(knots = knots_version,
                                       degree = degree_version,
                                       r_linear = tail_version,
                                       l_linear = tail_version,
                                       knots_type = "rel_freq",
                                       include_first_basis = True))
]

model = SFMAModel(model_data, variables)
model.attach(data)

model.fit(verbose = True, outlier_pct = pct_trim,
          eta_options = {"method": "bounded", "bounds": [0.0, 100.0]}
          )

# see beta coefficients
model.get_beta_dict()

# calculate inefficiency scores
data['log_ineff_pkg'] = model.get_inefficiency()
data['ineff_pkg'] = np.exp(model.get_inefficiency())

# make predictions with mean covariate values
Y_hat_null = model.predict(data)

# add preedicted columns to dataframe: 
data['log_pred_null'] = -Y_hat_null
data['val_pred'] = np.exp(data['log_pred_null'])

# calculate absolute inefficiency
data['abs_ineff'] = data['val'] - data['val_pred']

# save outliers for plotting
data['outliers'] = model.weights == 0.0

# order DF by SDI for correct plotting
data = data.sort_values(x)

# create outlier filter (after sorting to make sure index is correct)
outliers = data['outliers'] == True

log_knots = (variables[0].spline.knots)

print((log_knots))

# save out data frame with frontier predictions, inefficiency scores
data.to_csv(f"{output_dir}{cause_label}_{version}_{measure_name}_inefficiencies.csv", index=False)
